OPENEUROPA-2251: Refactor after rebase.
