+++ modules/oe_translation_poetry/tests/Functional/PoetryAddLanguageTest.php
@@ -5,8 +5,6 @@ declare(strict_types = 1);
 namespace Drupal\Tests\oe_translation_poetry\Functional;
 
 use Drupal\node\NodeInterface;
-use Drupal\oe_translation_poetry\Plugin\tmgmt\Translator\PoetryTranslator;
-use Drupal\tmgmt\Entity\Job;
 
 /**
  * Tests adding languages to requests made to Poetry.
@@ -81,23 +79,7 @@ class PoetryAddLanguageTest extends PoetryTranslationTestBase {
     // Send the translations for each job.
     $jobs = $this->jobStorage->loadMultiple();
     $this->notifyWithDummyTranslations($jobs);
-
-    $this->jobStorage->resetCache();
-    $this->entityTypeManager->getStorage('tmgmt_job_item')->resetCache();
-    $jobs = $this->jobStorage->loadMultiple();
-    foreach ($jobs as $job) {
-      $this->assertEqual($job->getState(), Job::STATE_ACTIVE);
-      $this->assertEqual($job->get('poetry_state')->value, PoetryTranslator::POETRY_STATUS_TRANSLATED);
-
-      $items = $job->getItems();
-      $item = reset($items);
-      $data = $this->container->get('tmgmt.data')->filterTranslatable($item->getData());
-      foreach ($data as $field => $info) {
-        $this->assertNotEmpty($info['#translation']);
-        $this->assertEqual($info['#translation']['#text'], $info['#text'] . ' - ' . $job->getTargetLangcode());
-      }
-    }
-
+    $this->assertJobsAreTranslated();
   }
 
   /**
